+++ portal_frontend/next.config.ts
@@ -2,7 +2,9 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
+  output: "export",
   images: {
+    unoptimized: true,
     remotePatterns: [
       {
         protocol: 'http',
