+++ portal_frontend/src/pages/favicon.ico
