+++ portal_frontend/src/pages/login/page.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 "use client";
 
-import useStore from "@/store";
+import useStore from "@/src/store";
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
 
@@ -10,11 +10,9 @@ export default function Login() {
   const setAccessToken = useStore((state) => state.setAccessToken)
   const router = useRouter();
 
-  let web: any = null;
-
   useEffect(() => {
     import("@privacybydesign/yivi-frontend").then((yivi: any) => {
-      web = yivi.newWeb({
+      const web = yivi.newWeb({
         debugging: false,            // Enable to get helpful output in the browser console
         element:   '#yivi-web-form', // Which DOM element to render to
       
@@ -53,7 +51,7 @@ export default function Login() {
           alert(err);
         });
     });
-  }, []);
+  }, [router, setAccessToken]);
 
   return (
     <div className="flex justify-center items-center">
